All paint logic was moved to separate composable function
